+++ src/views/figure_editable.ts
@@ -20,11 +20,7 @@ import {
   FileManagement,
   UnsupportedFormatFileIcon,
 } from '@manuscripts/style-guide'
-import {
-  encode,
-  ManuscriptEditorView,
-  ManuscriptNode,
-} from '@manuscripts/transform'
+import { ManuscriptEditorView, ManuscriptNode } from '@manuscripts/transform'
 import { createElement } from 'react'
 import ReactDOM from 'react-dom'
 
@@ -32,11 +28,11 @@ import {
   FigureOptions,
   FigureOptionsProps,
 } from '../components/views/FigureDropdown'
+import { buildFileMap } from '../lib/build-file-models'
 import { getActualAttrs } from '../lib/track-changes-utils'
 import { createEditableNodeView } from './creators'
 import { EditableBlockProps } from './editable_block'
 import { FigureView } from './figure'
-import { buildLightManuscript } from './figure_element'
 import { figureUploader } from './figure_uploader'
 import ReactSubView from './ReactSubView'
 
@@ -183,7 +179,7 @@ export class FigureEditableView extends FigureView<
       const componentProps: FigureOptionsProps = {
         can,
         files,
-        filesMap: encode(buildLightManuscript(this.view.state.doc)),
+        filesMap: buildFileMap(this.view.state.doc),
         handleDownload,
         handleUpload,
         handleDetach,
